fix: make sidebars active when modal is open
